Split actors per target type
